refactor: 예외 처리 클래스 패키지 세분화
