+++ src/main/java/roomescape/common/exception/handler/GlobalExceptionHandler.java
@@ -1,14 +1,23 @@
-package roomescape.common.exceptionHandler;
+package roomescape.common.exception.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.common.exceptionHandler.dto.ExceptionResponse;
+import roomescape.common.exception.handler.dto.ExceptionResponse;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
+
+    @ExceptionHandler(value = IllegalArgumentException.class)
+    public ResponseEntity<ExceptionResponse> illegalArgument(final IllegalArgumentException exception, final HttpServletRequest request) {
+        ExceptionResponse exceptionResponse = new ExceptionResponse(
+                400, "[ERROR] " + exception.getMessage(), request.getRequestURI()
+        );
+
+        return ResponseEntity.badRequest().body(exceptionResponse);
+    }
 
     @ExceptionHandler(value = HttpMessageNotReadableException.class)
     public ResponseEntity<ExceptionResponse> notReadable(
