+++ src/main/java/roomescape/common/exception/handler/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package roomescape.common.exceptionHandler.dto;
+package roomescape.common.exception.handler.dto;
 
 public record ExceptionResponse(int status, String message, String path) {
 }
