+++ src/test/java/roomescape/theme/ThemeIntegrationTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.common.exceptionHandler.dto.ExceptionResponse;
+import roomescape.common.exception.handler.dto.ExceptionResponse;
 import roomescape.theme.dto.ThemeResponse;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
